dracut-module/password-agent.c: Skip flaky test on non-AMD64
It is difficult to find a suitable buffer size large enough to
predictably make send() fail with EMSGSIZE on a PF_LOCAL socket, while
still being small enough to reliably malloc() twice on any platform
and in any environment.  The existing size was experimentally
determined on AMD64; skip the test for all other platforms.

* dracut-module/password-agent.c
  (test_send_password_to_socket_EMSGSIZE): Skip this test on non-AMD64
  platforms.

Closes: 953799
Reported-by: Michael Biebl <biebl@debian.org>