Do not hardcode directory names; get them from pkg-config
Similarly to the directory for systemd unit files (output of
"pkg-config systemd --variable=systemdsystemunitdir"), the values of
the "tmpfilesdir" and "sysusersdir" variables should not be hardcoded
in debian/mandos.dirs and debian/mandos-client.dirs.  Fix this the
same way as with systemdsystemunitdir; i.e. remove the directories
from the .dirs files, and instead create the directories in override
rules in debian/rules.

* debian/mandos-client.dirs (usr/lib/sysusers.d): Removed.
* debian/mandos.dirs (usr/lib/tmpfiles.d, usr/lib/sysusers.d):
  Removed.
* debian/rules (override_dh_installdirs-indep): Also create
  directories for the "tmpfilesdir" and "sysusersdir" variables from
  pkg-config.
  (override_dh_installdirs-arch): New; create directory for the
  "sysusersdir" variable from pkg-config.