Override lintian experimental tag "executable-in-usr-lib"
Lintian reports executable files in /usr/lib, since binaries which are
executed internally are supposed to be stored in /usr/libexec.  But
the executables we store in /usr/lib are all files to be copied into
the initial RAM disk image, and are never executed from their storage
location in /usr/lib.  Except one; the Dracut module-setup.sh file.
But that file seems to be required to be executable; the
module-setup.sh file of every other Dracut module is also an
executable file.

* debian/mandos-client.lintian-overrides: Override experimental tag
  "executable-in-usr-lib" about files which are not actually executed
  from this location, and only exist to be copied to the initial RAM
  disk image.  Also the Dracut module-setup.sh file, which merely
  follows the pattern of all other Dracut module-setup.sh files from
  other Dracut modules.